+++ kview/src/main/java/dev/ikm/komet/kview/mvvm/view/changeset/ExportController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.ikm.komet.kview.mvvm.view.export;
+package dev.ikm.komet.kview.mvvm.view.changeset;
 
 import dev.ikm.komet.framework.events.EvtBus;
 import dev.ikm.komet.framework.events.EvtBusFactory;
