+++ kview/src/main/java/module-info.java
@@ -120,8 +120,6 @@ module dev.ikm.komet.kview {
 
     exports dev.ikm.komet.kview.mvvm.view.navigation;
     opens dev.ikm.komet.kview.mvvm.view.navigation to javafx.fxml, org.carlfx.cognitive;
-    exports dev.ikm.komet.kview.mvvm.view.export;
-    opens dev.ikm.komet.kview.mvvm.view.export;
 
     provides KometNodeFactory with DetailsNodeFactory, PropertiesNodeFactory;
